# api-spec.yaml
openapi: 3.0.3
info:
  title: FinanceOS API
  description: Production-grade personal finance backend API
  version: 1.0.0
  contact:
    name: FinanceOS
servers:
  - url: http://localhost:6969
    description: Local development server
...
paths:
...
  /api/v1/counterparties:
    get:
      tags: [Counterparties]
      summary: List counterparties
      operationId: listCounterparties
      security:
        - sessionAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paged list of counterparties
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCounterpartyResponse'
    post:
      tags: [Counterparties]
      summary: Create counterparty
      operationId: createCounterparty
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCounterpartyRequest'
      responses:
        '201':
          description: Counterparty created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterpartyResponse'
  /api/v1/counterparties/{id}:
    put:
      tags: [Counterparties]
      summary: Update counterparty
      operationId: updateCounterparty
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCounterpartyRequest'
      responses:
        '200':
          description: Counterparty updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterpartyResponse'
    delete:
      tags: [Counterparties]
      summary: Delete counterparty
      operationId: deleteCounterparty
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Counterparty deleted
  /api/v1/lendings:
    get:
      tags: [Lendings]
      summary: List lendings
      operationId: listLendings
      security:
        - sessionAuth: []
      parameters:
        - name: counterpartyId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/LendingStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paged list of lendings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageLendingResponse'
    post:
      tags: [Lendings]
      summary: Create lending
      operationId: createLending
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLendingRequest'
      responses:
        '201':
          description: Lending created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingResponse'
  /api/v1/lendings/{id}:
    get:
      tags: [Lendings]
      summary: Get lending detail
      operationId: getLendingDetail
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lending detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingResponse'
    put:
      tags: [Lendings]
      summary: Update lending
      operationId: updateLending
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLendingRequest'
      responses:
        '200':
          description: Lending updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingResponse'
    delete:
      tags: [Lendings]
      summary: Delete lending
      operationId: deleteLending
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Lending deleted
  /api/v1/lendings/{id}/repayments:
    post:
      tags: [Lendings]
      summary: Add lending repayment
      operationId: addLendingRepayment
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLendingRepaymentRequest'
      responses:
        '201':
          description: Repayment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendingRepaymentResponse'
  /api/v1/lendings/{id}/repayments/{repaymentId}:
    delete:
      tags: [Lendings]
      summary: Delete lending repayment
      operationId: deleteLendingRepayment
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: repaymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Repayment deleted
  /api/v1/lendings/{id}/write-off:
    post:
      tags: [Lendings]
      summary: Write off lending
      operationId: writeOffLending
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lending written off
  /api/v1/lendings/{id}/reopen:
    post:
      tags: [Lendings]
      summary: Reopen lending
      operationId: reopenLending
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lending reopened
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: FINANCEOS_SESSION
  schemas:
...
    LendingDirection:
      type: string
      enum: [lent, borrowed]
    LendingStatus:
      type: string
      enum: [outstanding, partially_repaid, settled, written_off]
...
    CreateCounterpartyRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        notes: { type: string }
    UpdateCounterpartyRequest:
      type: object
      properties:
        name: { type: string }
        notes: { type: string }
    CounterpartyResponse:
      type: object
      required: [id, name, lentOutstanding, borrowedOutstanding, netPosition, openLendingCount, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        notes: { type: string }
        lentOutstanding: { type: number }
        borrowedOutstanding: { type: number }
        netPosition: { type: number }
        openLendingCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateLendingRequest:
      type: object
      required: [direction, amount, lendDate]
      properties:
        counterpartyId: { type: string, format: uuid }
        newCounterpartyName: { type: string }
        direction: { $ref: '#/components/schemas/LendingDirection' }
        amount: { type: number }
        lendDate: { type: string, format: date }
        expectedReturnDate: { type: string, format: date }
        transactionId: { type: string, format: uuid }
        notes: { type: string }
    UpdateLendingRequest:
      type: object
      properties:
        direction: { $ref: '#/components/schemas/LendingDirection' }
        amount: { type: number }
        lendDate: { type: string, format: date }
        expectedReturnDate: { type: string, format: date }
        notes: { type: string }
    LendingResponse:
      type: object
      required: [id, counterparty, direction, amount, outstanding, lendDate, status, repayments, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        counterparty: { $ref: '#/components/schemas/CounterpartyResponse' }
        direction: { $ref: '#/components/schemas/LendingDirection' }
        amount: { type: number }
        outstanding: { type: number }
        lendDate: { type: string, format: date }
        expectedReturnDate: { type: string, format: date }
        status: { $ref: '#/components/schemas/LendingStatus' }
        notes: { type: string }
        transactionId: { type: string, format: uuid }
        repayments: { type: array, items: { $ref: '#/components/schemas/LendingRepaymentResponse' } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateLendingRepaymentRequest:
      type: object
      required: [amount, date]
      properties:
        amount: { type: number }
        date: { type: string, format: date }
        transactionId: { type: string, format: uuid }
    LendingRepaymentResponse:
      type: object
      required: [id, lendingId, amount, date, createdAt]
      properties:
        id: { type: string, format: uuid }
        lendingId: { type: string, format: uuid }
        amount: { type: number }
        date: { type: string, format: date }
        transactionId: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
...
    PageCounterpartyResponse:
      type: object
      required: [content, number, size, totalElements, totalPages, first, last, empty]
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/CounterpartyResponse' } }
        number: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }
    PageLendingResponse:
      type: object
      required: [content, number, size, totalElements, totalPages, first, last, empty]
      properties:
        content: { type: array, items: { $ref: '#/components/schemas/LendingResponse' } }
        number: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }